requirements.txt
asgiref==3.3.1
astroid==2.4.2
boto3==1.17.41
botocore==1.20.41
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cors-headers==3.7.0
django-multipleselectfield==0.2.0.post2
djangorestframework==3.12.2
idna==2.10
isort==5.7.0
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.6
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2020.5
requests==2.25.1
s3transfer==0.3.6
six==1.15.0
spotipy==2.17.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.3
wrapt==1.12.1
youtube-dl==2021.4.1
